More tests for AuthStore and refactoring
