+++ metrics/web/test/test_utils/project_metrics_store_mock.dart
@@ -1,0 +1,5 @@
+import 'package:metrics/features/dashboard/presentation/state/project_metrics_store.dart';
+import 'package:mockito/mockito.dart';
+
+/// Mock implementation of the [ProjectMetricsStore].
+class ProjectMetricsStoreMock extends Mock implements ProjectMetricsStore {}
